+++ src/main/java/com/symphony/devsol/federation/activity/DeleteConnectContact.java
@@ -1,12 +1,12 @@
 package com.symphony.devsol.federation.activity;
 
 import com.symphony.bdk.workflow.swadl.v1.activity.BaseActivity;
+
 import lombok.Data;
 import lombok.EqualsAndHashCode;
-import java.math.BigDecimal;
 
 @Data
 @EqualsAndHashCode(callSuper=true)
 public class DeleteConnectContact extends BaseActivity {
-  private BigDecimal userId;
+  private long userId;
 }
